chore: 공업사 패키지 구조 개선 작업 #84
