+++ src/main/java/com/carumuch/capstone/bidding/presentation/dto/BidResDto.java
@@ -1,7 +1,7 @@
 package com.carumuch.capstone.bidding.presentation.dto;
 
 import com.carumuch.capstone.bidding.domain.Bid;
-import com.carumuch.capstone.bodyshop.dto.BodyShopInfoResDto;
+import com.carumuch.capstone.bodyshop.presentation.dto.BodyShopInfoResDto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
