+++ src/main/java/com/carumuch/capstone/bodyshop/application/BodyShopService.java
@@ -1,18 +1,18 @@
-package com.carumuch.capstone.bodyshop.service;
+package com.carumuch.capstone.bodyshop.application;
 
 import com.carumuch.capstone.bidding.domain.Bid;
-import com.carumuch.capstone.bodyshop.model.BodyShop;
+import com.carumuch.capstone.bodyshop.domain.BodyShop;
 import com.carumuch.capstone.bidding.domain.BidStatus;
-import com.carumuch.capstone.bodyshop.dto.BodyShopBidCreateReqDto;
-import com.carumuch.capstone.bodyshop.dto.BodyShopBidDetailResDto;
-import com.carumuch.capstone.bodyshop.dto.BodyShopBidPageResDto;
-import com.carumuch.capstone.bodyshop.dto.BodyShopBidUpdateReqDto;
-import com.carumuch.capstone.bodyshop.dto.BodyShopInfoResDto;
-import com.carumuch.capstone.bodyshop.dto.BodyShopPageResDto;
-import com.carumuch.capstone.bodyshop.dto.BodyShopRegistrationReqDto;
-import com.carumuch.capstone.bodyshop.dto.BodyShopUpdateReqDto;
+import com.carumuch.capstone.bodyshop.presentation.dto.BodyShopBidCreateReqDto;
+import com.carumuch.capstone.bodyshop.presentation.dto.BodyShopBidDetailResDto;
+import com.carumuch.capstone.bodyshop.presentation.dto.BodyShopBidPageResDto;
+import com.carumuch.capstone.bodyshop.presentation.dto.BodyShopBidUpdateReqDto;
+import com.carumuch.capstone.bodyshop.presentation.dto.BodyShopInfoResDto;
+import com.carumuch.capstone.bodyshop.presentation.dto.BodyShopPageResDto;
+import com.carumuch.capstone.bodyshop.presentation.dto.BodyShopRegistrationReqDto;
+import com.carumuch.capstone.bodyshop.presentation.dto.BodyShopUpdateReqDto;
 import com.carumuch.capstone.bidding.domain.BidRepository;
-import com.carumuch.capstone.bodyshop.repository.BodyShopRepository;
+import com.carumuch.capstone.bodyshop.domain.BodyShopRepository;
 import com.carumuch.capstone.common.legacy.exception.ErrorCode;
 import com.carumuch.capstone.common.legacy.exception.CustomException;
 import com.carumuch.capstone.user.domain.User;
