+++ src/main/java/com/carumuch/capstone/bodyshop/presentation/dto/BodyShopBidDetailResDto.java
@@ -1,4 +1,4 @@
-package com.carumuch.capstone.bodyshop.dto;
+package com.carumuch.capstone.bodyshop.presentation.dto;
 
 
 import io.swagger.v3.oas.annotations.media.Schema;
