+++ src/main/java/com/carumuch/capstone/bodyshop/presentation/dto/BodyShopBidUpdateReqDto.java
@@ -1,4 +1,4 @@
-package com.carumuch.capstone.bodyshop.dto;
+package com.carumuch.capstone.bodyshop.presentation.dto;
 
 
 import com.carumuch.capstone.common.legacy.validation.ValidationGroups;
