+++ src/main/java/com/carumuch/capstone/bodyshop/presentation/dto/BodyShopRegistrationReqDto.java
@@ -1,6 +1,6 @@
-package com.carumuch.capstone.bodyshop.dto;
+package com.carumuch.capstone.bodyshop.presentation.dto;
 
-import com.carumuch.capstone.bodyshop.model.Location;
+import com.carumuch.capstone.bodyshop.domain.Location;
 import com.carumuch.capstone.common.legacy.validation.ValidationGroups;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
@@ -10,18 +10,19 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-@Schema(name = "BodyShopUpdateReqDto: 공업사 수정 요청 Dto")
-public class BodyShopUpdateReqDto {
+@Schema(name = "BodyShopRegistrationReqDto: 공업사 등록 요청 Dto")
+public class BodyShopRegistrationReqDto {
+
     @NotBlank(message = "공업사 이름이 입력되지 않았습니다.", groups = ValidationGroups.NotBlankGroup.class)
     @Schema(description = "공업사 이름 입니다.",
-            example = "카우머치수리")
+            example = "차박고카센터")
     private String name;
 
     private Location location;
 
     @NotBlank(message = "공업사 설명이 입력되지 않았습니다.", groups = ValidationGroups.NotBlankGroup.class)
     @Schema(description = "공업사 설명 입니다. (최대 100자)",
-            example = "홍대사거리에서 홍대동, 홍대방면으로 직진하시면 홍대교회 건너편에 위치하고 있습니다." +
+            example = "고양사거리에서 고양동, 고양방면으로 직진하시면 고양교회 건너편에 위치하고 있습니다." +
                     "찾아오기 힘드시다면 바로 전화해주세요! 친철하게 안내 해드리겠습니다!")
     private String description;
 
@@ -35,16 +36,16 @@ public class BodyShopUpdateReqDto {
     private String link;
 
     @Schema(description = "픽업 여부 입니다.",
-            example = "false")
+            example = "true")
     private boolean pickupAvailability;
 
     @Builder
-    public BodyShopUpdateReqDto(String name, Location location, String description, String phoneNumber, String link, boolean pickupAvailability) {
+    public BodyShopRegistrationReqDto(String name, Location location, String description, String link, String phoneNumber, boolean pickupAvailability) {
         this.name = name;
         this.location = location;
         this.description = description;
-        this.phoneNumber = phoneNumber;
         this.link = link;
+        this.phoneNumber = phoneNumber;
         this.pickupAvailability = pickupAvailability;
     }
 }
